add condition before deleting dag reference

+++ centric_tools/task_management/dag_manager/handler.py
@@ -16,6 +16,11 @@ class DagWriter:
     def delete_dag(self, filename: str, dag_id: str, delay: int = 5, client: IDagClient = None):
         if delay:
             time.sleep(delay)
-        self.storage.delete(filename)
-        if client:
+        delete_success = self.storage.delete(filename)
+        if delete_success and client:
+            # NOTE: We only delete the dag reference if the file
+            # is deleted successfully because, if the dag reference is deleted and the
+            # file still exists in the dag bag the client will have no knowledge of the dag because the reference
+            # was deleted and treat the dag as new and rerun it again which wil result in an infinite loop of dag reference deletion and
+            # dag rerun
             client.delete_dag(dag_id)
